refactor(event cache): move `PaginationToken` from the pagination to room/mod file
