+++ crates/matrix-sdk/src/event_cache/pagination.rs
@@ -280,6 +280,29 @@ impl RoomPagination {
     }
 }
 
+/// Pagination token data, indicating in which state is the current pagination.
+#[derive(Clone, Debug)]
+pub(super) enum PaginationToken {
+    /// We never had a pagination token, so we'll start back-paginating from the
+    /// end, or forward-paginating from the start.
+    None,
+    /// We paginated once before, and we received a prev/next batch token that
+    /// we may reuse for the next query.
+    HasMore(String),
+    /// We've hit one end of the timeline (either the start or the actual end),
+    /// so there's no need to continue paginating.
+    HitEnd,
+}
+
+impl From<Option<String>> for PaginationToken {
+    fn from(token: Option<String>) -> Self {
+        match token {
+            Some(val) => Self::HasMore(val),
+            None => Self::None,
+        }
+    }
+}
+
 /// A type representing whether the timeline has been reset.
 #[derive(Debug)]
 pub enum TimelineHasBeenResetWhilePaginating {
